fix: ventas con tipo de pago
